Elimino archivos innecesarios y renombro archivos de test
