+++ app/test/test_e2e/test_event_avg_rating.py
