+++ app/test/test_integration/test_event.py
@@ -5,7 +5,7 @@ from django.test import Client, TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from app.models import Event, User, Rating
+from app.models import Event, User
 
 
 class BaseEventTestCase(TestCase):
@@ -68,8 +68,8 @@ class EventsListViewTest(BaseEventTestCase):
 
         # Verificar que los eventos están ordenados por fecha
         events = list(response.context["events"])
-        self.assertEqual(events[0].id, self.event1.id)
-        self.assertEqual(events[1].id, self.event2.id)
+        self.assertEqual(events[0].id, self.event1.id) # type: ignore
+        self.assertEqual(events[1].id, self.event2.id) # type: ignore
 
     def test_events_view_with_organizer_login(self):
         """Test que verifica que la vista events funciona cuando el usuario es organizador"""
@@ -90,7 +90,7 @@ class EventsListViewTest(BaseEventTestCase):
 
         # Verificar que redirecciona al login
         self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
+        self.assertTrue(response.url.startswith("/accounts/login/")) # type: ignore
 
 
 class EventDetailViewTest(BaseEventTestCase):
@@ -102,22 +102,22 @@ class EventDetailViewTest(BaseEventTestCase):
         self.client.login(username="regular", password="password123")
 
         # Hacer petición a la vista event_detail
-        response = self.client.get(reverse("event_detail", args=[self.event1.id]))
+        response = self.client.get(reverse("event_detail", args=[self.event1.id])) # type: ignore
 
         # Verificar respuesta
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, "app/event_detail.html")
         self.assertIn("event", response.context)
-        self.assertEqual(response.context["event"].id, self.event1.id)
+        self.assertEqual(response.context["event"].id, self.event1.id) # type: ignore
 
     def test_event_detail_view_without_login(self):
         """Test que verifica que la vista event_detail redirige a login cuando el usuario no está logueado"""
         # Hacer petición a la vista event_detail sin login
-        response = self.client.get(reverse("event_detail", args=[self.event1.id]))
+        response = self.client.get(reverse("event_detail", args=[self.event1.id])) # type: ignore
 
         # Verificar que redirecciona al login
         self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
+        self.assertTrue(response.url.startswith("/accounts/login/")) # type: ignore
 
     def test_event_detail_view_with_invalid_id(self):
         """Test que verifica que la vista event_detail devuelve 404 cuando el evento no existe"""
@@ -158,7 +158,7 @@ class EventFormViewTest(BaseEventTestCase):
 
         # Verificar que redirecciona a events
         self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse("events"))
+        self.assertEqual(response.url, reverse("events")) # type: ignore
 
     def test_event_form_view_without_login(self):
         """Test que verifica que la vista event_form redirige a login cuando el usuario no está logueado"""
@@ -167,7 +167,7 @@ class EventFormViewTest(BaseEventTestCase):
 
         # Verificar que redirecciona al login
         self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
+        self.assertTrue(response.url.startswith("/accounts/login/")) # type: ignore
 
     def test_event_form_edit_existing(self):
         """Test que verifica que se puede editar un evento existente"""
@@ -175,12 +175,12 @@ class EventFormViewTest(BaseEventTestCase):
         self.client.login(username="organizador", password="password123")
 
         # Hacer petición a la vista event_form para editar evento existente
-        response = self.client.get(reverse("event_edit", args=[self.event1.id]))
+        response = self.client.get(reverse("event_edit", args=[self.event1.id])) # type: ignore
 
         # Verificar respuesta
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, "app/event_form.html")
-        self.assertEqual(response.context["event"].id, self.event1.id)
+        self.assertEqual(response.context["event"].id, self.event1.id) # type: ignore
 
 
 class EventFormSubmissionTest(BaseEventTestCase):
@@ -204,7 +204,7 @@ class EventFormSubmissionTest(BaseEventTestCase):
 
         # Verificar que redirecciona a events
         self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse("events"))
+        self.assertEqual(response.url, reverse("events")) # type: ignore
 
         # Verificar que se creó el evento
         self.assertTrue(Event.objects.filter(title="Nuevo Evento").exists())
@@ -231,11 +231,11 @@ class EventFormSubmissionTest(BaseEventTestCase):
         }
 
         # Hacer petición POST para editar el evento
-        response = self.client.post(reverse("event_edit", args=[self.event1.id]), updated_data)
+        response = self.client.post(reverse("event_edit", args=[self.event1.id]), updated_data) # type: ignore
 
         # Verificar que redirecciona a events
         self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse("events"))
+        self.assertEqual(response.url, reverse("events")) # type: ignore
 
         # Verificar que el evento fue actualizado
         self.event1.refresh_from_db()
@@ -258,16 +258,16 @@ class EventDeleteViewTest(BaseEventTestCase):
         self.client.login(username="organizador", password="password123")
 
         # Verificar que el evento existe antes de eliminar
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
         # Hacer una petición POST para eliminar el evento
-        response = self.client.post(reverse("event_delete", args=[self.event1.id]))
+        response = self.client.post(reverse("event_delete", args=[self.event1.id])) # type: ignore
 
         # Verificar que redirecciona a la página de eventos
         self.assertRedirects(response, reverse("events"))
 
         # Verificar que el evento ya no existe
-        self.assertFalse(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertFalse(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
     def test_event_delete_with_regular_user(self):
         """Test que verifica que un usuario regular no puede eliminar un evento"""
@@ -275,16 +275,16 @@ class EventDeleteViewTest(BaseEventTestCase):
         self.client.login(username="regular", password="password123")
 
         # Verificar que el evento existe antes de intentar eliminarlo
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
         # Hacer una petición POST para intentar eliminar el evento
-        response = self.client.post(reverse("event_delete", args=[self.event1.id]))
+        response = self.client.post(reverse("event_delete", args=[self.event1.id])) # type: ignore
 
         # Verificar que redirecciona a la página de eventos sin eliminar
         self.assertRedirects(response, reverse("events"))
 
         # Verificar que el evento sigue existiendo
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
     def test_event_delete_with_get_request(self):
         """Test que verifica que la vista redirecciona si se usa GET en lugar de POST"""
@@ -292,13 +292,13 @@ class EventDeleteViewTest(BaseEventTestCase):
         self.client.login(username="organizador", password="password123")
 
         # Hacer una petición GET para intentar eliminar el evento
-        response = self.client.get(reverse("event_delete", args=[self.event1.id]))
+        response = self.client.get(reverse("event_delete", args=[self.event1.id])) # type: ignore
 
         # Verificar que redirecciona a la página de eventos
         self.assertRedirects(response, reverse("events"))
 
         # Verificar que el evento sigue existiendo
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
     def test_event_delete_nonexistent_event(self):
         """Test que verifica el comportamiento al intentar eliminar un evento inexistente"""
@@ -320,58 +320,16 @@ class EventDeleteViewTest(BaseEventTestCase):
     def test_event_delete_without_login(self):
         """Test que verifica que la vista redirecciona a login si el usuario no está autenticado"""
         # Verificar que el evento existe antes de intentar eliminarlo
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
         # Hacer una petición POST sin iniciar sesión
-        response = self.client.post(reverse("event_delete", args=[self.event1.id]))
+        response = self.client.post(reverse("event_delete", args=[self.event1.id])) # type: ignore
 
         # Verificar que redirecciona al login
         self.assertEqual(response.status_code, 302)
-        self.assertTrue(response.url.startswith("/accounts/login/"))
+        self.assertTrue(response.url.startswith("/accounts/login/")) # type: ignore
 
         # Verificar que el evento sigue existiendo
-        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists())
+        self.assertTrue(Event.objects.filter(pk=self.event1.id).exists()) # type: ignore
 
-class EventDetailIntegrationTest(TestCase):
-    def setUp(self):
-        # Usuario organizador
-         self.organizer = User.objects.create_user(
-            username="organizador_test",
-            email="organizador@example.com",
-            password="password123",
-            is_organizer=True,
-        )
-        # Usuario no organizador
-         self.regular_user = User.objects.create_user(
-            username="regular",
-            email="regular@test.com",
-            password="password123",
-            is_organizer=False,
-        )
-        # Crear algunos eventos de prueba
-         self.event1 = Event.objects.create(
-            title="Evento 1",
-            description="Descripción del evento 1",
-            scheduled_at=timezone.now() + datetime.timedelta(days=1),
-            organizer=self.organizer,
-        )
 
-         Rating.objects.create(usuario=self.regular_user, evento=self.event1, title="Buena", text="Me gustó", rating=4)
-         Rating.objects.create(usuario=self.regular_user, evento=self.event1, title="Excelente", text="Muy buena experiencia", rating=3)
-         Rating.objects.create(usuario=self.regular_user, evento=self.event1, title="Buena", text="Me gustó", rating=3)
-         Rating.objects.create(usuario=self.regular_user, evento=self.event1, title="Excelente", text="Muy buena experiencia", rating=1)
-
-    def test_event_detail_muestra_promedio_si_es_organizador(self):
-
-        self.client.login(username='organizador_test', password='password123')
-        response = self.client.get(reverse("event_detail", args=[self.event1.id]))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "2,75")  # el promedio esperado
-
-    def test_event_detail_no_muestra_promedio_si_no_es_organizador(self):
-        self.client.login(username='regular', password='password123')
-        response = self.client.get(reverse("event_detail", args=[self.event1.id]))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertNotContains(response, "2,75")
