+++ app/test/test_unit/test_event.py
@@ -3,7 +3,7 @@ import datetime
 from django.test import TestCase
 from django.utils import timezone
 
-from app.models import Event, User, Rating
+from app.models import Event, User
 
 
 class EventModelTest(TestCase):
@@ -81,7 +81,7 @@ class EventModelTest(TestCase):
         )
 
         self.assertFalse(success)
-        self.assertIn("title", errors)
+        self.assertIn("title", errors) # type: ignore
 
         # Verificar que no se creó ningún evento nuevo
         self.assertEqual(Event.objects.count(), initial_count)
@@ -141,67 +141,3 @@ class EventModelTest(TestCase):
         self.assertEqual(updated_event.description, new_description)
         self.assertEqual(updated_event.scheduled_at, original_scheduled_at)
 
-class EventRatingTestCase(TestCase):
-    def setUp(self):
-        self.organizer = User.objects.create_user(
-            username="organizador_test",
-            email="organizador@example.com",
-            password="password123",
-            is_organizer=True,
-        )
-
-        # Crear evento
-        self.event = Event.objects.create(
-            title="Evento de prueba",
-            description="Descripción del evento de prueba",
-            scheduled_at=timezone.now() + datetime.timedelta(days=1),
-            organizer=self.organizer,
-        )
-        self.regular_user = User.objects.create_user(
-            username="regular",
-            email="regular@test.com",
-            password="password123",
-            is_organizer=False,
-        )
-
-    def test_promedio_rating_sin_ratings(self):
-        promedio = self.event.promedio_rating()
-        self.assertEqual(promedio, 0)
-
-    def test_promedio_rating_con_ratings(self):
-        Rating.objects.create(
-            usuario=self.regular_user,
-            evento=self.event,
-            title="Buena",
-            text="Me gustó",
-            rating=4
-        )
-        Rating.objects.create(
-            usuario=self.regular_user,
-            evento=self.event,
-            title="Excelente",
-            text="Muy buena experiencia",
-            rating=5
-        )
-        Rating.objects.create(
-            usuario=self.regular_user,
-            evento=self.event,
-            title="Regular",
-            text="Podría mejorar",
-            rating=3
-        )
-
-        promedio = self.event.promedio_rating()
-        self.assertEqual(promedio, 4)  # (4 + 5 + 3) / 3 = 4.0
-
-    def test_promedio_rating_con_rating_unico(self):
-        Rating.objects.create(
-            usuario=self.regular_user,
-            evento=self.event,
-            title="Excelente",
-            text="Muy buena experiencia",
-            rating=5
-        )
-
-        promedio = self.event.promedio_rating()
-        self.assertEqual(promedio, 5)
